:truck: Move `lots_of_logging` to benchmark utils
