+++ tests/benchmarks/test_benchmarks.py
@@ -1,9 +1,10 @@
-from typing import Any, Generator
+from typing import Generator
 
 import pytest
 from pytest_benchmark.fixture import BenchmarkFixture
 
 import structlog_sentry_logger
+from tests.benchmarks import utils
 from tests.structlog_sentry_logger import test__config
 
 
@@ -27,19 +28,7 @@ def test_logging_stdlib_json_serializer(benchmark: BenchmarkFixture) -> None:
 
 def _benchmark_runner(benchmark: BenchmarkFixture, test_cases: dict) -> None:
     logger = structlog_sentry_logger.get_logger()
-    benchmark(lots_of_logging, logger=logger, test_cases=test_cases)
-
-
-def lots_of_logging(logger: Any, test_cases: dict) -> None:
-    for i in range(10):
-        for log_level_fn in [
-            logger.debug,
-            logger.info,
-            logger.warn,
-            logger.error,
-            logger.fatal,
-        ]:
-            log_level_fn(f"iter: {i}", **test_cases)
+    benchmark(utils.lots_of_logging, logger=logger, test_cases=test_cases)
 
 
 @pytest.fixture(scope="function")
