+++ tests/benchmarks/utils.py
@@ -1,0 +1,13 @@
+from typing import Any
+
+
+def lots_of_logging(logger: Any, test_cases: dict) -> None:
+    for i in range(10):
+        for log_level_fn in [
+            logger.debug,
+            logger.info,
+            logger.warn,
+            logger.error,
+            logger.fatal,
+        ]:
+            log_level_fn(f"iter: {i}", **test_cases)
